romapi: fix various misra issues
The comma operator should not be used. (misra-c2012-12.3)
A function should have a single point of exit at the end. (misra-c2012-15.5)
Both operands of an operator in which the usual arithmetic conversions
are performed shall have the same essential type category. (misra-c2012-10.4)
The precedence of operators within expressions should be made explicit. (misra-c2012-12.1)
The scope of some variables can be reduced. (variableScope)
Use the appropriate macros to get uint64_t low and high.

Signed-off-by: Viorel Suman <viorel.suman@nxp.com>